Core: move create_playthrough under Spoiler as method Core: split create_playthrough, allowing skipping of paths
